curtainctrl: Added simple proof of concept for BT module.
In each loop btInteractive() is called. If any character is received by
the bluetooth module, the function stays in a loop where it executes
command until an 'exit' (x or X) is received.
Two commands are implemented:
0: Increase RTC time with 1 hour
1: Decrease RTC time with 1 hour

Use e.g. the app 'Bluetooth terminal' to enter command.
Remember to connect phone and BT module.